refactor: Styled 폴더 -> Flex, LoadingSpinner 폴더로 분리
