+++ src/components/common/Button/index.jsx
@@ -1,19 +1,33 @@
-import React from 'react';
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
+import styled, { css } from 'styled-components';
 
-const Button = ({ children }) => {
-  return <StyledButton>{children}</StyledButton>;
-};
-
-export const StyledButton = styled.button`
+const Button = styled.button`
   border: none;
   background: none;
   cursor: pointer;
 `;
 
-Button.propTypes = {
-  children: PropTypes.element.isRequired,
-};
+export const CartButton = styled(Button)`
+  :hover {
+    svg path {
+      fill: ${({ theme }) => theme.colors.mint};
+    }
+  }
+`;
+
+export const CardDetailButton = styled(Button)`
+  ${({ theme }) => css`
+    width: 100%;
+    height: 60px;
+
+    font-size: 20px;
+    font-weight: 700;
+    background: ${theme.colors.gray};
+    color: ${theme.colors.white};
+
+    :hover {
+      background: ${theme.colors.mint};
+    }
+  `}
+`;
 
 export default Button;
