+++ src/components/common/Flex/index.jsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const FlexCenter = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const FlexSpaceBetween = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+
+export const FlexColumnCenter = styled(FlexCenter)`
+  flex-direction: column;
+`;
