+++ src/components/common/Header/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
+
 import { ReactComponent as CartIcon } from 'assets/icon/Cart.svg';
-import { css } from 'styled-components';
 
 const Header = () => {
   return (
